+++ services/web/server/src/simcore_service_webserver/projects/module_setup.py
@@ -17,7 +17,7 @@ from servicelib.aiohttp.rest_routing import (
 )
 
 from .._constants import APP_OPENAPI_SPECS_KEY
-from ..resources import resources
+from .._resources import resources
 from . import projects_handlers, projects_nodes_handlers, projects_tags_handlers
 from .config import CONFIG_SECTION_NAME, assert_valid_config
 from .projects_access import setup_projects_access
