+++ services/web/server/src/simcore_service_webserver/rest_utils.py
@@ -5,7 +5,7 @@ from typing import Optional
 import openapi_core
 import yaml
 from openapi_core.schema.specs.models import Spec as OpenApiSpecs
-from simcore_service_webserver.resources import resources
+from simcore_service_webserver._resources import resources
 
 from ._meta import api_version_prefix
