+++ services/web/server/tests/unit/conftest.py
@@ -17,7 +17,7 @@ from typing import Any, Callable, Dict, Iterable
 import pytest
 from openapi_core.schema.specs.models import Spec as OpenApiSpecs
 from pytest_simcore.helpers.utils_projects import empty_project_data
-from simcore_service_webserver.resources import resources
+from simcore_service_webserver._resources import resources
 from simcore_service_webserver.rest_utils import (
     get_openapi_specs_path,
     load_openapi_specs,
