+++ services/web/server/tests/unit/isolated/test_projects_utils.py
@@ -10,11 +10,11 @@ from typing import Any, Dict, Set
 import jsonschema
 import pytest
 from jsonschema import ValidationError
+from simcore_service_webserver._resources import resources
 from simcore_service_webserver.projects.projects_utils import (
     clone_project_document,
     project_get_depending_nodes,
 )
-from simcore_service_webserver.resources import resources
 
 
 def load_template_projects() -> Dict[str, Any]:
